refactor(conjugation): remove legacy planning shims
