+++ src/polyzymd/builders/conjugation/reactions/base.py
@@ -5,27 +5,6 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import Any, ClassVar
 
-from pydantic import BaseModel, Field
-
-
-class ReactionContext(BaseModel):
-    """Placeholder context passed to future reaction templates."""
-
-    model_config = {"arbitrary_types_allowed": True}
-
-    protein: Any | None = None
-    moiety: Any | None = None
-    metadata: dict[str, Any] = Field(default_factory=dict)
-
-
-class ReactionResult(BaseModel):
-    """Placeholder result returned by future reaction templates."""
-
-    model_config = {"arbitrary_types_allowed": True}
-
-    plan: Any | None = None
-    metadata: dict[str, Any] = Field(default_factory=dict)
-
 
 class ReactionTemplate(ABC):
     """Base class for conjugation reaction mechanisms.
@@ -72,20 +51,3 @@ class ReactionTemplate(ABC):
         Any
             A mechanism-owned generic resolved attachment plan.
         """
-
-    def plan(self, context: ReactionContext) -> ReactionResult:
-        """Resolve a generic plan from a reaction context."""
-        site_config = context.metadata.get("site_config") or context.metadata.get("site")
-        settings = context.metadata.get("settings")
-        if context.protein is None or context.moiety is None or site_config is None:
-            raise ValueError(
-                "Reaction planning requires context.protein, context.moiety, and "
-                "metadata['site_config']"
-            )
-        plan = self.resolve_attachment(
-            context.protein,
-            site_config,
-            context.moiety,
-            settings=settings,
-        )
-        return ReactionResult(plan=plan, metadata={"mechanism": self.name})
