+++ tests/test_conjugation_product_pablo.py
@@ -16,10 +16,8 @@ from polyzymd.builders.conjugation.pablo.product import (
     build_product_state_pablo_library,
     build_product_state_pablo_library_for_specs,
 )
-from polyzymd.builders.conjugation.polymer import (
-    GeneratedPolymerFragment,
-    generated_fragment_from_polymerist_pdb,
-)
+from polyzymd.builders.conjugation.polymer import GeneratedPolymerFragment
+from polyzymd.builders.conjugation.polymer.polymerist import generated_fragment_from_polymerist_pdb
 from polyzymd.builders.conjugation.structure.pdb import PdbAtomRecord
 from polyzymd.config.schema import ConjugationCcdPabloPolicyConfig
